requirements.txt
Flask-Dance[sqla]
alembic==1.6.4
click==8.0.1
coverage==5.5
flake8==3.9.2
Flask==2.0.1
Flask-Cors==3.0.10
Flask-Migrate==3.0.0
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.8.1
greenlet==1.1.0
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
mccabe==0.6.1
mysqlclient==2.0.3
pycodestyle==2.7.0
bcrypt
pyjwt
flask-bcrypt
pyflakes==2.3.1
python-dateutil==2.8.1
python-editor==1.0.4
six==1.16.0
SQLAlchemy==1.4.15
Werkzeug==2.0.1
sendgrid==6.1.0
premailer==3.6.1
boto3